requirements.txt
pillow
aiogram == 3
paramiko
sqlite3